docs: fix Search component for algolia search
