--- apps/docs/src/components/Search.jsx
@@ -1,76 +1,0 @@
-import { useCallback, useEffect, useState } from 'react';
-import { createPortal } from 'react-dom';
-import { DocSearchModal, useDocSearchKeyboardEvents } from '@docsearch/react';
-
-const docSearchConfig = {
-  appId: import.meta.env.NEXT_PUBLIC_DOCSEARCH_APP_ID,
-  apiKey: import.meta.env.NEXT_PUBLIC_DOCSEARCH_API_KEY,
-  indexName: import.meta.env.NEXT_PUBLIC_DOCSEARCH_INDEX_NAME,
-};
-
-function Hit({ hit, children }) {
-  return <a href={hit.url}>{children}</a>;
-}
-
-function SearchIcon(props) {
-  return (
-    <svg aria-hidden='true' viewBox='0 0 20 20' {...props}>
-      <path d='M16.293 17.707a1 1 0 0 0 1.414-1.414l-1.414 1.414ZM9 14a5 5 0 0 1-5-5H2a7 7 0 0 0 7 7v-2ZM4 9a5 5 0 0 1 5-5V2a7 7 0 0 0-7 7h2Zm5-5a5 5 0 0 1 5 5h2a7 7 0 0 0-7-7v2Zm8.707 12.293-3.757-3.757-1.414 1.414 3.757 3.757 1.414-1.414ZM14 9a4.98 4.98 0 0 1-1.464 3.536l1.414 1.414A6.98 6.98 0 0 0 16 9h-2Zm-1.464 3.536A4.98 4.98 0 0 1 9 14v2a6.98 6.98 0 0 0 4.95-2.05l-1.414-1.414Z' />
-    </svg>
-  );
-}
-
-export function Search() {
-  let [isOpen, setIsOpen] = useState(false);
-  let [modifierKey, setModifierKey] = useState();
-
-  const onOpen = useCallback(() => {
-    setIsOpen(true);
-  }, [setIsOpen]);
-
-  const onClose = useCallback(() => {
-    setIsOpen(false);
-  }, [setIsOpen]);
-
-  useDocSearchKeyboardEvents({ isOpen, onOpen, onClose });
-
-  useEffect(() => {
-    setModifierKey(/(Mac|iPhone|iPod|iPad)/i.test(navigator.platform) ? '⌘' : 'Ctrl ');
-  }, []);
-
-  return (
-    <>
-      <button
-        type='button'
-        className='group flex h-6 w-6 items-center justify-center sm:justify-start md:h-auto md:w-80 md:flex-none md:rounded-lg md:py-2.5 md:pl-4 md:pr-3.5 md:text-sm md:ring-1 md:ring-slate-200 md:hover:ring-slate-300 dark:md:bg-slate-800/75 dark:md:ring-inset dark:md:ring-white/5 dark:md:hover:bg-slate-700/40 dark:md:hover:ring-slate-500 lg:w-96'
-        onClick={onOpen}
-      >
-        <SearchIcon className='h-5 w-5 flex-none fill-slate-400 group-hover:fill-slate-500 dark:fill-slate-500 md:group-hover:fill-slate-400' />
-        <span className='sr-only md:not-sr-only md:ml-2 md:text-slate-500 md:dark:text-slate-400'>
-          Search docs
-        </span>
-        {modifierKey && (
-          <kbd className='ml-auto hidden font-medium text-slate-400 dark:text-slate-500 md:block'>
-            <kbd className='font-sans'>{modifierKey}</kbd>
-            <kbd className='font-sans'>K</kbd>
-          </kbd>
-        )}
-      </button>
-      {isOpen &&
-        createPortal(
-          <DocSearchModal
-            {...docSearchConfig}
-            initialScrollY={window.scrollY}
-            onClose={onClose}
-            hitComponent={Hit}
-            navigator={{
-              navigate({ itemUrl }) {
-                history.push(itemUrl);
-              },
-            }}
-          />,
-          document.body,
-        )}
-    </>
-  );
-}
